requirements.txt
# Core
Django==4.2.16
djangorestframework==3.15.2
django-environ==0.11.2

# Database
psycopg2-binary==2.9.9

# Auth & OAuth
django-allauth==0.63.3
social-auth-app-django==5.4.1

# Media & Documents
Pillow==10.4.0
reportlab==4.2.2

# AI
openai==1.40.0

# API
django-cors-headers==4.4.0
drf-spectacular==0.27.2
django-filter==24.3

# Cache & Queue
django-redis==5.4.0
redis==5.0.8
celery==5.4.0
django-celery-beat==2.6.0
django-celery-results==2.5.1

# Static files
whitenoise==6.7.0

# Production server
gunicorn==22.0.0
